수욜 네번째

+++ FlexDelivery/src/main/java/com/project/fd/admin/temporary/model/AdminTemporaryService.java
@@ -8,12 +8,12 @@ import com.project.fd.owner.store.model.OwnerStoresVO;
 
 public interface AdminTemporaryService {
 	public List<AdminTemporaryVO> editList();
-	public AdminTemporaryVO editDetail(long no);
+	public AdminTemporaryVO editDetail(int no);
 	public int editAgree(AdminTemporaryVO vo);
 	/*
 	public int editOwner(OwnerVO ownerVo);
 	public int editOwnerRegister(OwnerRegisterVO registerVo);
 	public int editStores(OwnerStoresVO storesVo);
 	*/
-	public int editDeny(long no);
+	public int editDeny(int no);
 }
